feat:Implementing the shopping cart

+++ src/main/java/com/buyology/backend/model/CartItem.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Entity
 @Data
 @AllArgsConstructor
@@ -25,8 +27,8 @@ public class CartItem {
     private Product product;
 
     private Integer quantity;
-    private Double discount;
-    private Double productPrice;
+    private BigDecimal discount;
+    private BigDecimal productPrice;
 
 
 }
